backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
python-dotenv==1.0.1
httpx==0.28.1
openai==1.57.4
greenlet==3.1.1
python-dateutil==2.9.0.post0
